(refactor) reogranize loader files and update deploy
