+++ .eslintignore
@@ -1,1 +1,1 @@
-src/loader.js
+src/s1/loader.js
