+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+<a name="0.0.1"></a>
+
+## [0.0.1](https://github.com/openmail/system1-cmp/compare/v0.0.0...v0.0.1) (2018-07-02)
+
+### Features
+
+ - [x] Initial Release!
+ - [x] Versioned Deployment and updated README for installing CMP
