+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "appnexus-cmp",
-  "version": "0.0.3",
+  "version": "0.0.1",
   "scripts": {
     "clean": "rm -rf dist",
     "dev": "cross-env NODE_ENV=development webpack-dev-server --inline --hot --progress --config config/webpack.config.babel.js",
@@ -8,13 +8,13 @@
     "start": "serve dist -s -c 1",
     "prestart": "npm run build",
     "build": "rimraf ./dist && yarn build:original && yarn build:s1",
-    "deploy": "yarn deploy:original && yarn deploy:s1",
+    "deploy": "rimraf ./dist  && yarn deploy:original && yarn deploy:s1",
     "build:original": "cross-env NODE_ENV=production webpack --progress --config config/webpack.config.babel.js",
     "deploy:original": "yarn build:original && yarn upload:original",
-    "upload:original": "cross-var s3-deploy './dist/{*.?(js|html),docs/**}' --cwd './dist' --region us-west-2 --bucket s1-layout-cdn/christian-test --gzip --cache 1440 --invalidate '/christian-test/*.js /christian-test/*.html /christian-test/docs/**'",
+    "upload:original": "cross-var s3-deploy './dist/{*.?(js|html),docs/**}' --cwd './dist' --region us-west-2 --bucket s1-layout-cdn/cmp --gzip --cache 1440 --invalidate '/cmp/*.js /cmp/* /cmp/docs/*'",
     "build:s1": "cross-env NODE_ENV=production webpack --progress --config config/s1.webpack.config.babel.js",
     "deploy:s1": "yarn build:s1 && yarn upload:s1",
-    "upload:s1": "cross-var s3-deploy './dist/$npm_package_version/**' --cwd './dist/$npm_package_version' --region us-west-2 --bucket s1-layout-cdn/christian-test/$npm_package_version --gzip --preventUpdates --immutable --ext js",
+    "upload:s1": "cross-var s3-deploy './dist/$npm_package_version/**' --cwd './dist/$npm_package_version' --region us-west-2 --bucket s1-layout-cdn/cmp/$npm_package_version --gzip --preventUpdates --immutable --ext js",
     "prebuild": "npm run clean && mkdirp dist",
     "test": "npm run -s lint && jest --coverage",
     "test:watch": "npm run -s test -- --watch",
