+++ src/lib/vendor.js
@@ -3,8 +3,6 @@ import 'whatwg-fetch';
 import config from './config';
 import log from './log';
 
-const PUB_VENDOR_LOCATION = '/.well-known/pubvendors.json';
-
 /**
  * Fetch the pubvendors.json from the local domain
  */
