+++ src/s1/cmp.js
@@ -6,7 +6,7 @@ import 'core-js/fn/array/find';
 import 'core-js/fn/array/map';
 import 'core-js/fn/object/keys';
 
-import cmp from '../loader';
+import cmp from './loader';
 import {init} from '../lib/init';
 import log from '../lib/log';
 import {readCookie, writeCookie} from "../lib/cookie/cookie";
