--- src/s1/constants.js
@@ -1,31 +1,0 @@
-export const GDPRCountries = [
-	'AT',
-	'BE',
-	'BG',
-	'HR',
-	'CY',
-	'CZ',
-	'DK',
-	'EE',
-	'FI',
-	'FR',
-	'DE',
-	'GR',
-	'HU',
-	'IE',
-	'IT',
-	'LV',
-	'LT',
-	'LU',
-	'MT',
-	'NL',
-	'PL',
-	'PT',
-	'RO',
-	'SK',
-	'SI',
-	'ES',
-	'SE',
-	'GB',
-	'US' //  @FIXME @christian remove
-];
